todo crud işlemleri user bazlı  düzenlendi. Ufak fonksiyon düzenlemeleri yapıldı.

+++ 1-start-todos/repositories/todosRepository.js
@@ -1,12 +1,12 @@
 import Todo from '../models/todosModel.js';
 
-const getNextTodoId = async () => {
-    const lastTodo = await Todo.findOne().sort({ id: -1 });
+const _getNextTodoId = async (userId) => {
+    const lastTodo = await Todo.findOne({ userId: userId }).sort({ id: -1 });
     return lastTodo ? lastTodo.id + 1 : 1;
 };
 
 const createTodo = async (title, completed, userId) => {
-    const nextId = await getNextTodoId();
+    const nextId = await _getNextTodoId(userId);
     const newTodo = new Todo();
     newTodo.id = nextId;
     newTodo.title = title;
@@ -16,23 +16,23 @@ const createTodo = async (title, completed, userId) => {
     return newTodo;
 }
 
-const deleteTodo = async (id) => {
-    const deletedTodo = await Todo.findOneAndDelete({ id });
+const deleteTodo = async (id, userId) => {
+    const deletedTodo = await Todo.findOneAndDelete({ id: id, userId: userId });
     return deletedTodo;
 };
 
-const updateTodo = async (id, title, completed) => {
-    const updatedTodo = await Todo.findOneAndUpdate({ id }, { title, completed }, { new: true });
+const updateTodo = async (id, title, completed, userId) => {
+    const updatedTodo = await Todo.findOneAndUpdate({ id: id, userId: userId }, { title, completed }, { new: true });
     return updatedTodo;
 };
 
-const getAllTodos = async () => {
-    const todos = await Todo.find({});
+const getAllTodos = async (userId) => {
+    const todos = await Todo.find({ userId: userId });
     return todos;
 };
 
-const getTodoById = async (id) => {
-    const todo = await Todo.findOne({ id });
+const getTodoById = async (id, userId) => {
+    const todo = await Todo.findOne({ id: id, userId: userId });
     return todo;
 };
